% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_inputs}
\alias{create_inputs}
\title{Use selectInput, textInput, dateInput based on column name}
\usage{
create_inputs(INPUT, VALSCSV, colname)
}
\arguments{
\item{INPUT}{session input}

\item{colname}{character string to match colname}

\item{VALS}{session values}
}
\value{
*Input for shiny
}
\description{
Use selectInput, textInput, dateInput based on column name
}
