% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modal_fill_pdf}
\alias{modal_fill_pdf}
\title{Modal from which to fill PDF}
\usage{
modal_fill_pdf(INPUT, VALSCSV)
}
\arguments{
\item{INPUT}{session input}

\item{VALS}{data reactiveVal}
}
\description{
Modal from which to fill PDF
}
