% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{obsev_FILENAMES}
\alias{obsev_FILENAMES}
\title{Take filename INPUT and parse}
\usage{
obsev_FILENAMES(INPUT, VALS)
}
\arguments{
\item{INPUT}{session input}

\item{VALS}{data reactiveVal}
}
\description{
Take filename INPUT and parse
}
