% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{obsev_go_fill_pdf}
\alias{obsev_go_fill_pdf}
\title{Use selectInput, textInput, dateInput based on column name}
\usage{
obsev_go_fill_pdf(INPUT, VALS)
}
\arguments{
\item{INPUT}{session input}

\item{VALS}{session values}
}
\value{
*Input for shiny
}
\description{
Use selectInput, textInput, dateInput based on column name
}
