% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{obsev_go_get_pdf}
\alias{obsev_go_get_pdf}
\title{Allow use to donwload the PDF made and save to tmp}
\usage{
obsev_go_get_pdf(INPUT, OUTPUT, VALS)
}
\arguments{
\item{INPUT}{session input}

\item{OUTPUT}{session output}

\item{VALS}{session values}
}
\description{
Allow use to donwload the PDF made and save to tmp
}
