% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{obsev_manual}
\alias{obsev_manual}
\title{Take filename INPUT and parse}
\usage{
obsev_manual(INPUT, VALS)
}
\arguments{
\item{INPUT}{session input}

\item{VALS}{data reactiveVal}
}
\description{
Take filename INPUT and parse
}
