% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_name_vec}
\alias{split_name_vec}
\title{Split into the many millions of single chars on this vec}
\usage{
split_name_vec()
}
\description{
Split into the many millions of single chars on this vec
}
